+++ service-monitor/src/main/java/com/yahoo/vespa/service/duper/ControllerApplication.java
@@ -1,7 +1,6 @@
 // Copyright 2018 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.service.duper;
 
-import com.yahoo.component.Version;
 import com.yahoo.config.provision.ClusterSpec;
 import com.yahoo.config.provision.NodeType;
 import com.yahoo.vespa.applicationmodel.ServiceType;
@@ -15,15 +14,6 @@ public class ControllerApplication extends ConfigServerLikeApplication {
 
     public ControllerApplication() {
         super("controller", NodeType.controller, ClusterSpec.Type.container, ServiceType.CONTROLLER);
-    }
-
-    @Override
-    public ClusterSpec getClusterSpecWithVersion(Version version) {
-        return ClusterSpec.request(getClusterSpecType(), getClusterSpecId())
-                          .vespaVersion(version)
-                          .stateful(true)
-                          .exclusive(true)
-                          .build();
     }
 
 }
